Lineを描くViewを単一ファイルに切り出した。
